fix: remove @emotion/core types after running yarn
There is a known issue with storybook that the @emotion/core dependency augments
the React `HTMLAttributes` with an additional `css` property. This causes the
compiled types of react-bootstrap-typeahead to include that attribute as well,
which in turns breaks Typescript compilation of projects using it.

See: https://github.com/emotion-js/emotion/issues/1257#issuecomment-547188357