Add keepOpen prop to <Typeahead>
The keepOpen property can be used to prevent default behaviour of
closing the menu and resetting the components internal state when
selecting an item in multi mode.

This can be for example be used to implement "select multiple when
ctrl is pressed down" -functionality. For example when your users
often have the need to select every item with a specific prefix.